Build native php

+++ sapi/src/builder/library/php.php
@@ -5,22 +5,23 @@ use SwooleCli\Preprocessor;
 
 return function (Preprocessor $p) {
     $php_install_prefix = BUILD_PHP_INSTALL_PREFIX;
+    $php_tag = BUILD_PHP_VERSION;
     $php_src = $p->getPhpSrcDir();
     $build_dir = $p->getBuildDir();
 
     $p->addLibrary(
         (new Library('php'))
-            ->withUrl('https://github.com/php/php-src/archive/refs/tags/php-' . BUILD_PHP_VERSION . '.tar.gz')
+            ->withUrl('https://github.com/php/php-src/archive/refs/tags/php-' . $php_tag . '.tar.gz')
             ->withHomePage('https://www.php.net/')
             ->withLicense('https://github.com/php/php-src/blob/master/LICENSE', Library::LICENSE_PHP)
             ->withPrefix($php_install_prefix)
-            ->withCleanBuildDirectory()
+            ->withBuildCached(false)
+            ->withInstallCached(false)
             ->withBuildScript(
                 <<<EOF
-            if test -d {$php_src} ; then
-                rm -rf {$php_src}
-            fi
             cd {$build_dir}/
+            test -d {$php_src} && rm -rf {$php_src}
+
             cp -rf php {$php_src}
             rm -rf php
             cd {$build_dir}/
@@ -28,6 +29,5 @@ EOF
             )
             ->withLdflags('')
             ->withPkgConfig('')
-            ->withBuildCached(false)
     );
 };
